+++ internal/aws/cwlogs/pusher_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package awsemfexporter
+package cwlogs
 
 import (
 	"fmt"
@@ -51,10 +51,10 @@ func TestConcurrentPushAndFlush(t *testing.T) {
 	for i := 0; i < concurrency; i++ {
 		go func(ii int) {
 			for j := 0; j < 10; j++ {
-				emfPusher.addLogEntry(newLogEvent(current, fmt.Sprintf("batch-%d-%d", ii, j)))
+				emfPusher.AddLogEntry(NewEvent(current, fmt.Sprintf("batch-%d-%d", ii, j)))
 			}
 			time.Sleep(1000 * time.Millisecond)
-			emfPusher.forceFlush()
+			emfPusher.ForceFlush()
 			wg.Done()
 		}(i)
 	}
@@ -64,7 +64,7 @@ func TestConcurrentPushAndFlush(t *testing.T) {
 	maxEventPayloadBytes = defaultMaxEventPayloadBytes
 }
 
-func newMockPusherWithEventCheck(check func(msg string)) (pusher, string) {
+func newMockPusherWithEventCheck(check func(msg string)) (Pusher, string) {
 	logger := zap.NewNop()
 	tmpfolder, _ := ioutil.TempDir("", "")
 	svc := newAlwaysPassMockLogClient(func(args mock.Arguments) {
@@ -83,7 +83,7 @@ func newMockPusherWithEventCheck(check func(msg string)) (pusher, string) {
 //
 func TestLogEvent_eventPayloadBytes(t *testing.T) {
 	testMessage := "test message"
-	logEvent := newLogEvent(0, testMessage)
+	logEvent := NewEvent(0, testMessage)
 	assert.Equal(t, len(testMessage)+perEventHeaderBytes, logEvent.eventPayloadBytes())
 }
 
@@ -91,37 +91,37 @@ func TestValidateLogEventWithMutating(t *testing.T) {
 	maxEventPayloadBytes = 64
 
 	logger := zap.NewNop()
-	logEvent := newLogEvent(0, "abcdefghijklmnopqrstuvwxyz0123456789abcdefghijklmnopqrstuvwxyz0123456789abcdefghijklmnopqrstuvwxyz0123456789")
-	logEvent.logGeneratedTime = time.Now()
+	logEvent := NewEvent(0, "abcdefghijklmnopqrstuvwxyz0123456789abcdefghijklmnopqrstuvwxyz0123456789abcdefghijklmnopqrstuvwxyz0123456789")
+	logEvent.GeneratedTime = time.Now()
 	err := logEvent.Validate(logger)
 	assert.Nil(t, err)
-	assert.True(t, *logEvent.inputLogEvent.Timestamp > int64(0))
-	assert.Equal(t, 64-perEventHeaderBytes, len(*logEvent.inputLogEvent.Message))
+	assert.True(t, *logEvent.InputLogEvent.Timestamp > int64(0))
+	assert.Equal(t, 64-perEventHeaderBytes, len(*logEvent.InputLogEvent.Message))
 
 	maxEventPayloadBytes = defaultMaxEventPayloadBytes
 }
 
 func TestValidateLogEventFailed(t *testing.T) {
 	logger := zap.NewNop()
-	logEvent := newLogEvent(0, "")
+	logEvent := NewEvent(0, "")
 	err := logEvent.Validate(logger)
 	assert.NotNil(t, err)
 	assert.Equal(t, "empty log event message", err.Error())
 
 	invalidTimestamp := time.Now().AddDate(0, -1, 0)
-	logEvent = newLogEvent(invalidTimestamp.Unix()*1e3, "test")
+	logEvent = NewEvent(invalidTimestamp.Unix()*1e3, "test")
 	err = logEvent.Validate(logger)
 	assert.NotNil(t, err)
 	assert.Equal(t, "the log entry's timestamp is older than 14 days or more than 2 hours in the future", err.Error())
 }
 
 //
-//  logEventBatch Tests
+//  eventBatch Tests
 //
 func TestLogEventBatch_timestampWithin24Hours(t *testing.T) {
 	min := time.Date(2017, time.June, 20, 23, 38, 0, 0, time.Local)
 	max := min.Add(23 * time.Hour)
-	logEventBatch := &logEventBatch{
+	logEventBatch := &eventBatch{
 		maxTimestampMs: max.UnixNano() / 1e6,
 		minTimestampMs: min.UnixNano() / 1e6,
 	}
@@ -147,17 +147,17 @@ func TestLogEventBatch_timestampWithin24Hours(t *testing.T) {
 
 func TestLogEventBatch_sortLogEvents(t *testing.T) {
 	totalEvents := 10
-	logEventBatch := &logEventBatch{
+	logEventBatch := &eventBatch{
 		putLogEventsInput: &cloudwatchlogs.PutLogEventsInput{
 			LogEvents: make([]*cloudwatchlogs.InputLogEvent, 0, totalEvents)}}
 
 	for i := 0; i < totalEvents; i++ {
 		timestamp := rand.Int()
-		logEvent := newLogEvent(
+		logEvent := NewEvent(
 			int64(timestamp),
 			fmt.Sprintf("message%v", timestamp))
 		fmt.Printf("logEvents[%d].Timestamp=%d.\n", i, timestamp)
-		logEventBatch.putLogEventsInput.LogEvents = append(logEventBatch.putLogEventsInput.LogEvents, logEvent.inputLogEvent)
+		logEventBatch.putLogEventsInput.LogEvents = append(logEventBatch.putLogEventsInput.LogEvents, logEvent.InputLogEvent)
 	}
 
 	logEventBatch.sortLogEvents()
@@ -193,7 +193,7 @@ func TestPusher_newLogEventBatch(t *testing.T) {
 	p, tmpFolder := newMockPusher()
 	defer os.RemoveAll(tmpFolder)
 
-	logEventBatch := newLogEventBatch(p.logGroupName, p.logStreamName)
+	logEventBatch := newEventBatch(p.logGroupName, p.logStreamName)
 	assert.Equal(t, int64(0), logEventBatch.maxTimestampMs)
 	assert.Equal(t, int64(0), logEventBatch.minTimestampMs)
 	assert.Equal(t, 0, logEventBatch.byteTotal)
@@ -208,10 +208,10 @@ func TestPusher_addLogEventBatch(t *testing.T) {
 	defer os.RemoveAll(tmpFolder)
 
 	cap := cap(p.logEventBatch.putLogEventsInput.LogEvents)
-	logEvent := newLogEvent(timestampMs, msg)
+	logEvent := NewEvent(timestampMs, msg)
 
 	for i := 0; i < cap; i++ {
-		p.logEventBatch.putLogEventsInput.LogEvents = append(p.logEventBatch.putLogEventsInput.LogEvents, logEvent.inputLogEvent)
+		p.logEventBatch.putLogEventsInput.LogEvents = append(p.logEventBatch.putLogEventsInput.LogEvents, logEvent.InputLogEvent)
 	}
 
 	assert.Equal(t, cap, len(p.logEventBatch.putLogEventsInput.LogEvents))
@@ -225,7 +225,7 @@ func TestPusher_addLogEventBatch(t *testing.T) {
 	assert.Equal(t, 1, len(p.logEventBatch.putLogEventsInput.LogEvents))
 
 	p.logEventBatch.minTimestampMs, p.logEventBatch.maxTimestampMs = timestampMs, timestampMs
-	assert.NotNil(t, p.addLogEvent(newLogEvent(timestampMs+(time.Hour*24+time.Millisecond*1).Nanoseconds()/1e6, msg)))
+	assert.NotNil(t, p.addLogEvent(NewEvent(timestampMs+(time.Hour*24+time.Millisecond*1).Nanoseconds()/1e6, msg)))
 	assert.Equal(t, 1, len(p.logEventBatch.putLogEventsInput.LogEvents))
 
 	assert.Nil(t, p.addLogEvent(nil))
@@ -245,12 +245,12 @@ func TestAddLogEventWithValidation(t *testing.T) {
 	defer os.RemoveAll(tmpFolder)
 	largeEventContent := strings.Repeat("a", defaultMaxEventPayloadBytes)
 
-	logEvent := newLogEvent(timestampMs, largeEventContent)
-	expectedTruncatedContent := (*logEvent.inputLogEvent.Message)[0:(defaultMaxEventPayloadBytes-perEventHeaderBytes-len(truncatedSuffix))] + truncatedSuffix
+	logEvent := NewEvent(timestampMs, largeEventContent)
+	expectedTruncatedContent := (*logEvent.InputLogEvent.Message)[0:(defaultMaxEventPayloadBytes-perEventHeaderBytes-len(truncatedSuffix))] + truncatedSuffix
 
-	p.addLogEntry(logEvent)
-	assert.Equal(t, expectedTruncatedContent, *logEvent.inputLogEvent.Message)
+	p.AddLogEntry(logEvent)
+	assert.Equal(t, expectedTruncatedContent, *logEvent.InputLogEvent.Message)
 
-	logEvent = newLogEvent(timestampMs, "")
+	logEvent = NewEvent(timestampMs, "")
 	assert.NotNil(t, p.addLogEvent(logEvent))
 }